Builds and Runs!
The problem was a module import error. I needed to call package_name.module_name instead of just module_name for all my own modules in this package. I'd like to see if I could just import STIM_Module so I don't need the package name on each of my modules

IMPORTANT
File paths need to be forward slash and not backward slashes. This is fixed in Sandy-Dev which will be my next merge commit

IMPORTANT
Temporary directories need to be checked and or made